+++ src/res/dataLoaderCompat.cjs
@@ -1,0 +1,5 @@
+"use strict";
+
+const importedDataLoader = require("dataloader");
+
+exports.DataLoader = importedDataLoader.default ?? importedDataLoader;